pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spectrum-orchestration-env"
version = "0.1.0"
description = "A Gymnasium-compatible RL environment for 5G/6G dynamic spectrum management, built on Meta's OpenEnv framework."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
keywords = ["reinforcement-learning", "gymnasium", "openenv", "spectrum-management", "5g", "pytorch"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "openenv-core[core] @ git+https://github.com/meta-pytorch/OpenEnv.git@v0.2.3",
    "fastmcp",
    "gymnasium>=0.29.0",
    "torch>=2.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "openai>=1.3.0",
]

[project.optional-dependencies]
train = ["matplotlib>=3.7.0"]
dev = ["pytest>=7.0.0"]

[project.scripts]
server = "server.app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "server*"]
